1214: Move RemovedStoreContent to a separate component
